// interfaces/companies.ts
export interface ICompanies {
    Id: number;
    Company: string;
    RFC: string;
    Subsidiary: string;
    Turn: string;
    LegalProxy: string;
    State: string;
    Town: string;
    Address: string;
    Type: string;
    Charter: string;
    Zone: string;
    VAT: number;
    CompanyType_Id: number;
    CompanyType: string;
    CreatedAt: Date;
    TaxRegime: string;
    Phone: string;

}